+++ config/splat.ric.yaml
@@ -31,10 +31,13 @@ segments:
       - [0x19C70, rodata]
       - [0x19EE0, rodata]
       - [0x1A454, rodata]
+      - [0x1A788, .rodata, 1F348]
+      - [0x1A7A8, rodata]
       - [0x1AB04, rodata]
       - [0x1AB68, .rodata, 32324]
       - [0x1AB88, rodata]
       - [0x1AC60, c, 1AC60]
+      - [0x1F348, c, 1F348]
       - [0x32324, c, 32324]
       - [0x37B60, data]
   - [0x39A58]
