+++ include/sfx.h
@@ -42,6 +42,7 @@
 #define NA_SE_PL_BT_FIREBALL 0x680
 #define NA_SE_PL_MAX_HP_MP_INCREASED 0x687
 #define NA_SE_EV_CLOCK_TICK 0x6A1
+#define NA_SE_PL_RIC_UNK_6E2 0x6E2
 #define NA_SE_VO_AL_PUNCH 0x6F1
 #define NA_SE_AL_BACKSLIDE 0x617
 #define NA_SE_VO_AL_WHAT 0x6F2
