+++ src/dra/4768C.c
@@ -1248,7 +1248,7 @@ void func_800F2404(s32 arg0) {
         ptr -= 1;
     }
 
-    D_80097420 = 0;
+    D_80097420[0] = 0;
     D_80097424 = 0;
     D_80097448 = 0;
     D_8009744C = 0;
